move status sign back to where it was
