+++ Movecraft/src/main/java/net/countercraft/movecraft/Movecraft.java
@@ -27,6 +27,7 @@ import net.countercraft.movecraft.features.contacts.ContactsCommand;
 import net.countercraft.movecraft.features.contacts.ContactsManager;
 import net.countercraft.movecraft.features.contacts.ContactsSign;
 import net.countercraft.movecraft.features.status.StatusManager;
+import net.countercraft.movecraft.features.status.StatusSign;
 import net.countercraft.movecraft.listener.*;
 import net.countercraft.movecraft.localisation.I18nSupport;
 import net.countercraft.movecraft.mapUpdater.MapUpdateManager;
